requirements.txt
streamlit==1.45.1
pandas==2.1.0
aiohttp==3.9.1
requests==2.31.0
nltk==3.8.1
transformers==4.36.2
torch==2.3.1
tqdm==4.66.1
google-generativeai==0.8.5
python-dotenv==1.1.0
numpy==1.26.4
nest-asyncio==1.6.0
wikipedia==1.4.0
aiohappyeyeballs==2.6.1
huggingface-hub==0.31.2 